+++ Sources/Pathos/permissions.swift
@@ -10,6 +10,7 @@ import Darwin
 /// - Parameter path: The path for which the permissions are get.
 /// - Throws: System error encountered while attempting to read permissions at the path.
 /// - SeeAlso: To work with `Path` or `PathRepresentable`, use `PathRepresentable.permissions`.
+@available(*, deprecated, message: "use `metadata(atPath:followSymbol:).permissions` instead")
 public func permissions(forPath path: String) throws -> FilePermission {
     var status = stat()
     if lstat(path, &status) != 0 {
@@ -27,7 +28,7 @@ public func permissions(forPath path: String) throws -> FilePermission {
 /// - Throws: System error encountered while attempting to read or change permissions at the path.
 /// - SeeAlso: To work with `Path` or `PathRepresentable`, use `PathRepresentable.add(_:)`.
 public func add(_ permissions: FilePermission, forPath path: String) throws {
-    let existingPermission = try permissions(forPath:)(path)
+    let existingPermission = try metadata(atPath: path).permissions
     try set(existingPermission.union(permissions), forPath: path)
 }
 
@@ -40,7 +41,7 @@ public func add(_ permissions: FilePermission, forPath path: String) throws {
 /// - Throws: System error encountered while attempting to read or change permissions at the path.
 /// - SeeAlso: To work with `Path` or `PathRepresentable`, use `PathRepresentable.remove(_:)`.
 public func remove(_ permissions: FilePermission, forPath path: String) throws {
-    let existingPermission = try permissions(forPath:)(path)
+    let existingPermission = try metadata(atPath: path).permissions
     try set(existingPermission.subtracting(permissions), forPath: path)
 }
 
@@ -64,6 +65,7 @@ extension PathRepresentable {
     /// If an error is encountered while reading the permission, `FilePermission(rawValue: 0)` will be the
     /// value. If an error is encountered while setting the permission, this will be a no-op.
     /// - SeeAlso: `permissions(forPath:)`.
+    @available(*, deprecated, message: "use `metadata(atPath:followSymbol:).permissions` and self.set(_:) instead")
     public var permissions: FilePermission {
         get {
             return (try? permissions(forPath:)(self.pathString)) ?? FilePermission(rawValue: 0)
@@ -71,6 +73,15 @@ extension PathRepresentable {
         set {
             try? set(_:forPath:)(newValue, self.pathString)
         }
+    }
+
+    /// Set permissions at a path. This function will *not* follow symbolic links, it'll change permissions for
+    /// the link itself instead.
+    ///
+    /// - Parameters permissions: Permissions to be set at the path.
+    /// - SeeAlso: `set(_:forPath:)`.
+    public func set(_ permissions: FilePermission) {
+        try? set(_:forPath:)(permissions, self.pathString)
     }
 
     /// Set additional permissions to existing permission at this path. If this is an symbolic link, the