+++ Tests/PathosTests/DefaultTemporaryDirectoryTests.swift
@@ -4,13 +4,13 @@ import XCTest
 final class DefaultTemporaryDirectoryTests: XCTestCase {
     func testDefaultsTemporaryDirectory() throws {
         XCTAssertTrue(exists(atPath: defaultTemporaryDirectory))
-        let permissions = try Pathos.permissions(forPath: defaultTemporaryDirectory)
+        let permissions = try metadata(atPath: defaultTemporaryDirectory).permissions
         XCTAssertTrue(permissions.contains([.ownerRead, .ownerWrite]))
     }
 
     func testPathRepresentableDefaultsTemporaryDirectory() throws {
         XCTAssertTrue(Path.defaultTemporaryDirectory.exists())
-        let permissions = Path.defaultTemporaryDirectory.permissions
+        let permissions = try XCTUnwrap(Path.defaultTemporaryDirectory.metadata()).permissions
         XCTAssertTrue(permissions.contains([.ownerRead, .ownerWrite]))
     }
 }