+++ Tests/PathosTests/PermissionsTests.swift
@@ -5,82 +5,82 @@ final class PermissionsTests: XCTestCase {
     func testReadingDefaultPermissions() throws {
         let destination = try createTemporaryFile()
         let expectedPermissions: FilePermission = [.ownerRead, .ownerWrite, .otherRead, .groupRead]
-        XCTAssertEqual(try permissions(forPath: destination), expectedPermissions)
+        XCTAssertEqual(try metadata(atPath: destination).permissions, expectedPermissions)
     }
 
     func testSettingPermissions() throws {
         let destination = try createTemporaryFile()
         let expectedPermissions: FilePermission = [.ownerRead, .ownerWrite, .otherRead, .groupRead]
-        XCTAssertEqual(try permissions(forPath: destination), expectedPermissions)
+        XCTAssertEqual(try metadata(atPath: destination).permissions, expectedPermissions)
         let resultPermissions: FilePermission = [.ownerRead, .ownerWrite]
 
         try set(resultPermissions, forPath: destination)
 
-        XCTAssertEqual(try permissions(forPath: destination), resultPermissions)
+        XCTAssertEqual(try metadata(atPath: destination).permissions, resultPermissions)
     }
 
     func testAddingPermissions() throws {
         let destination = try createTemporaryFile()
         let initialPermissions: FilePermission = [.ownerRead, .ownerWrite, .otherRead, .groupRead]
-        XCTAssertEqual(try permissions(forPath: destination), initialPermissions)
+        XCTAssertEqual(try metadata(atPath: destination).permissions, initialPermissions)
         let newPermission: FilePermission = .ownerExecute
         let resultPermissions: FilePermission = initialPermissions.union(newPermission)
 
         try Pathos.add(newPermission, forPath: destination)
 
-        XCTAssertEqual(try permissions(forPath: destination), resultPermissions)
+        XCTAssertEqual(try metadata(atPath: destination).permissions, resultPermissions)
     }
 
     func testRemovingPermissions() throws {
         let destination = try createTemporaryFile()
         let initialPermissions: FilePermission = [.ownerRead, .ownerWrite, .otherRead, .groupRead]
-        XCTAssertEqual(try permissions(forPath: destination), initialPermissions)
+        XCTAssertEqual(try metadata(atPath: destination).permissions, initialPermissions)
         let permissionToRemove: FilePermission = .ownerExecute
         let resultPermissions: FilePermission = initialPermissions.subtracting(permissionToRemove)
 
         try Pathos.remove(permissionToRemove, forPath: destination)
 
-        XCTAssertEqual(try permissions(forPath: destination), resultPermissions)
+        XCTAssertEqual(try metadata(atPath: destination).permissions, resultPermissions)
     }
 
     func testPathRepresentableReadingDefaultPermissions() {
         let destination = Path.createTemporaryFile()!
         let expectedPermissions: FilePermission = [.ownerRead, .ownerWrite, .otherRead, .groupRead]
-        XCTAssertEqual(destination.permissions, expectedPermissions)
+        XCTAssertEqual(destination.metadata()?.permissions, .some(expectedPermissions))
     }
 
     func testPathRepresentableSettingPermissions() throws {
-        var destination = Path.createTemporaryFile()!
+        let destination = Path.createTemporaryFile()!
         let expectedPermissions: FilePermission = [.ownerRead, .ownerWrite, .otherRead, .groupRead]
-        XCTAssertEqual(try permissions(forPath: destination.pathString), expectedPermissions)
+        XCTAssertEqual(try metadata(atPath: destination.pathString).permissions, expectedPermissions)
         let resultPermissions: FilePermission = [.ownerRead, .ownerWrite]
 
-        destination.permissions = resultPermissions
+        destination.set(resultPermissions)
 
-        XCTAssertEqual(destination.permissions, resultPermissions)
+        XCTAssertEqual(destination.metadata()?.permissions, .some(resultPermissions))
     }
 
     func testPathRepresentableAddingPermissions() throws {
         let destination = Path.createTemporaryFile()!
         let initialPermissions: FilePermission = [.ownerRead, .ownerWrite, .otherRead, .groupRead]
-        XCTAssertEqual(try permissions(forPath: destination.pathString), initialPermissions)
+        XCTAssertEqual(try metadata(atPath: destination.pathString).permissions, initialPermissions)
         let newPermission: FilePermission = .ownerExecute
         let resultPermissions: FilePermission = initialPermissions.union(newPermission)
 
         destination.add(newPermission)
 
-        XCTAssertEqual(destination.permissions, resultPermissions)
+        XCTAssertEqual(destination.metadata()?.permissions, .some(resultPermissions))
     }
 
     func testPathRepresentableRemovingPermissions() throws {
         let destination = Path.createTemporaryFile()!
         let initialPermissions: FilePermission = [.ownerRead, .ownerWrite, .otherRead, .groupRead]
-        XCTAssertEqual(try permissions(forPath: destination.pathString), initialPermissions)
+        XCTAssertEqual(try metadata(atPath: destination.pathString).permissions, initialPermissions)
         let permissionToRemove: FilePermission = .ownerExecute
         let resultPermissions: FilePermission = initialPermissions.subtracting(permissionToRemove)
 
         destination.remove(permissionToRemove)
 
-        XCTAssertEqual(destination.permissions, resultPermissions)
+        XCTAssertEqual(destination.metadata()?.permissions, .some(resultPermissions))
     }
 }